requirements.txt
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
torchvision>=0.11.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
fpdf>=1.7.2
pymysql>=1.0.2
python-dotenv>=0.19.0
scipy>=1.7.0
shap>=0.40.0
Pillow>=8.3.0
requests>=2.25.0
beautifulsoup4>=4.10.0
tqdm>=4.62.0 